Move ai-chat export into home feature
