+++ app/ai-chat/src/features/home/sidebar/conversation_item.rs
@@ -3,8 +3,7 @@ use crate::{
     components::{
         add_conversation::open_edit_conversation_dialog, delete_confirm::open_delete_confirm_dialog,
     },
-    export::ExportType,
-    features::home::{open_copy_conversation_dialog, open_export_conversation_prompt},
+    features::home::{ExportType, open_copy_conversation_dialog, open_export_conversation_prompt},
     foundation::assets::IconName,
     foundation::i18n::I18n,
     state::{ChatData, ChatDataEvent, WorkspaceStore},
